iprt/asm-amd64-x86.h: Corrected return type for non-inline ASMGetApicIdExt0B.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@86303 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/iprt/asm-amd64-x86.h
@@ -1389,7 +1389,7 @@ DECLINLINE(uint8_t) ASMGetApicId(void)
  * @returns the APIC ID.
  */
 #if RT_INLINE_ASM_EXTERNAL && RT_INLINE_ASM_USES_INTRIN < 16 /*?*/
-RT_ASM_DECL_PRAGMA_WATCOM(uint8_t) ASMGetApicIdExt0B(void);
+RT_ASM_DECL_PRAGMA_WATCOM(uint32_t) ASMGetApicIdExt0B(void);
 #else
 DECLINLINE(uint32_t) ASMGetApicIdExt0B(void)
 {
